requirements.txt
pandas==2.0.2
paho-mqtt==1.6.1
streamlit==1.27.2